+++ materialious/src/lib/i18n/locales/en.json
@@ -123,7 +123,8 @@
 			"proxyVideos": "Proxy videos",
 			"backgroundPlay": "Play audio in background",
 			"dash": "Dash",
-			"silenceSkipper": "Silence skipper (Experimental)"
+			"silenceSkipper": "Silence skipper (Experimental)",
+			"youtubeJsFallback": "Locally video processing fallback"
 		},
 		"bookmarklet": "Bookmarklet",
 		"instanceUrl": "Instance URL",
